Fix wrong assumption about the tablet pad strip position
This was a mistake when implementing support for tablet pad strips: the
position isn't integer! Or at least, not in the way we receive it.

libinput returns a clamped double with a range of [0, 1], but we
accidentally (and implicitly) converted it to an integer. And then,
Wayland clients assumedly only get either a 0 or a 1. This gets worse as
the Wayland protocol says the position is a range of [0, 65535].
Basically this never should have worked.

I fixed this by storing the intermediate double as a qreal in
TabletStripEvent, and then multiplying it by 65355 inside of the Wayland
interface like we do for other doubles.

+++ src/backends/libinput/events.h
@@ -407,7 +407,7 @@ class TabletPadStripEvent : public Event
 public:
     TabletPadStripEvent(libinput_event *event, libinput_event_type type);
 
-    int position() const
+    qreal position() const
     {
         return libinput_event_tablet_pad_get_strip_position(m_tabletPadEvent);
     }
