+++ src/input_event.h
@@ -244,7 +244,7 @@ struct TabletPadStripEvent
 {
     InputDevice *device;
     int number;
-    int position;
+    qreal position;
     bool isFinger;
     quint32 group;
     quint32 mode;
